Fix the scrolling problem in dynamic playlists.
After the track was changed the "scroll to current playing track" worked
as per user settings. But in case of dynamic playlists after the tracks
were added (automatically) the PrettyListView scrolled to the tracks added
(i.e. it scrolled to the bottom). It took me so long to figure this out
and even longer to find a solution; but the patch was only of 2 lines.

Anyway I got a better understanding of the Playlist Model-View architecture :)

BUG: 284214
FIXED-IN: 2.8
REVIEW: 111038