Always specify ColorGroup when fetching color from QPalette.
This commit fixes a huge number of color inconsistencies that existed
between the inital colors of widgets at Amarok startup, and the colors
that resulted from switching to the same theme at runtime.

It is *crucially important* when using QPalette::color() to always
specify the ColorGroup too, not just the ColorRole. Otherwise whatever
is the currently active group will be used.