Fix initial sizes, minimum sizes, and resizing of dock widgets.
Something that had annoyed me for a long time is when you resized
Amarok, the relative proportions of the three dock widgets would change.
Maybe I'm a bit OCD about UIs but it can't stand it when something is
almost symmetrical, but not quite. So that won't happen any more.

Also, the widgets now have sane minimum sizes.