Jamendo service: Use direct file download instead of BitTorrent.
When pressing the "Download" button in the Jamendo service, Amarok
would download a torrent file via deprecated API and then try to open
a torrent client for downloading the album. This was brittle and barely
ever worked as there are rarely any seeders and peers.

Now the download function uses the exact same code as the "Copy to
Collection" function, including the option to transcode on-the-fly.

BUG: 299434
DIGEST: Amarok now supports direct download of Jamendo albums instead
of downloading via BitTorrent. The download is now far more reliable, and
on-the-fly transcoding to other formats is possible.