Better handling of non-writeable UMS collections
If an UMS-collection is non-writeable (think freshly created ext4 on an USB-stick) a copy/move to
collection will silently fail. It can only be seen in the debug logs on the console.

The reason is that the UMS-collection is not actually checking if its writeable. Its just assuming
to be always writeable.

I check if the UMS-collection is actually writeable (mount point and if set the "music folder").

REVIEW: 111991