UMS: Fix setting album cover.
When Setting->Config->Metadata->Write covers to file is enabled, it
should be possible to write covers for albums located on an UMS
collection. The current status is that amarok does the write but
never updates itself to reflect the new cover. There is code to
achive this but is has been broken at some point.

File_p.h:358 tries to create a signal/slot connection to an invalid
slot. This is fixed by my patch.

REVIEW: 112255