Update anachronistic error message when plugins are not found.
The old message came in the form of an error dialog and gave some
advice that is no longer true for normal use cases. Usually plugins
are not found after upgrading Amarok, because ksyococa needs
updating. Due to a bug in KBuildSycocaProgressDialog this may
currently require an application restart, so we inform the user about
it.