MySqlEmbeddedStorage: fix failure to start embedded server because of wrong options
This was because we were passing --myisam-recover-options and this
options does not exist in MySQL 5.1 (and MariaDB). The *command-line*
option (as opposed to system variable name) to set recovery mode is
tricky:
MySQL <= 5.1: --myisam-recover [1]
MySQL >= 5.5: --myisam-recover-options, should also accept
              --myisam-recover [2]
MariaDB: --myisam-recover [3]

*However*, the *system variable name* has always been named
myisam_recover_options, that's why it worked in Amarok <= 2.8.

No need to backport this because of the above.

Simple fix is to rename the option to --myisam-recover

BUG: 323802
FIXED-IN: 2.9

[1] https://dev.mysql.com/doc/refman/5.1/en/myisam-start.html
[2] https://dev.mysql.com/doc/refman/5.5/en/myisam-start.html
[3] https://mariadb.com/kb/en/myisam-server-system-variables/