Fix sound glitch with "fade-out on pause" with GStreamer.
This is really just an ugly workaround for an issue with
(Phonon-)GStreamer. Essentially there is no way to flush the audio
buffers, so that implementing fade-out in combination with pausing
becomes a bit of a gamble.

Please test, as I cannot reproduce the sound glitch on my system!

BUG: 323729
FIXED-IN: 2.8.1
BACKPORT