OrganizeCollectionDialog: hide complex options behind a button
In commit 7c780bf5, Ralf addressed bug 283361 by only showing the preview
table when the screen is large enough, which I believe went too far.

This is a bit different approach, the file scheme and other options are
hidden behind the Options button when the dialog is shown, making it much
more cleaner. The rationale is that the user usually sets up the options
once, not every time (and beginner users may be happy with the default
schema right away).

BUG: 327201
FIXED-IN: 2.9