Make sure not to use video stream when transcoding to Opus.
I stumbled over a problem with transcoding to Opus. For example, if an
MP3 file contains cover art and you transcode it to Opus, ffmpeg creates
an Opus file with two streams: Ogg Theora (because of the picture) and
Opus. Thus, this patch makes sure not to use video streams.

REVIEW: 114563
FIXED-IN: 2.9