Fix infrequent Amarok freeze
If a SqlRegistry cleanup is in progress the SqlRegistry will first
lock it's own mutex and then the mutex of almost every SqlMeta object.

If at the same time one of the objects request tracks from SqlRegistry
then it will lock it's own mutex and then the SqlRegistry mutex.

Both are first locking their own mutex first and then trying to get
the other mutex, deadlocking.