Refactoring: Create StorageManager to provide SqlStorage
A implement a Storage manager that will provide the SqlStorage.

Reason: all parts of Amarok are relying on an SqlStorage.
Currently the collection manager does this job in addition to
it's main job of managing the collections.

Next step: move storage in separate plugin.

Note: deactivate all remainders of switching databases. That
never works and will hit a lot of Services since they suddenly are
without propperly initialized tables.