Use QCA to generate hashes for the Ampache login process
The sha256() function provided to generate hashes for Ampache is
incorrect (generates wrong hashes), and thus will not generate a proper
passphrase for Ampache versions greater than 35, which will fail with an
"incorrect password" message.

The solution is to use the hash functions provided by QCA, which work
correctly. However this adds a (conditional and optional) dependency to
QCA to build the Ampache plugin.