Fix: prevent potential issues with QString::arg
When using several "arg" one must make sure that there is not
a string inserted with %1 or else the next .arg calls would
also replace it.

Used one arg or just stream operations when usefull.

Was not even aware of this dangerous behavour of QString before
and it was also not mentioned in the documentation.