Do not set global shortcuts for the Media keys
Amarok used to grab the keys for:
* Qt::Key_MediaPlay
* Qt::Key_MediaStop
* Qt::Key_MediaNext
* Qt::Key_MediaPrevious

As those are global shortcuts they are taken away from any other
application when Amarok is running. E.g. if I watch a movie in VLC
and press the media play button (which VLC handles) Amarok will start
playback if it's running. If Amarok is not running, VLC will pause
the playback. This shows that a multi-media application should not
globally grab these keys.

In addition we just introduced a change [1] in plasma-workspace for the
Plasma/5.3 release to grab these shortcuts and bind them to the
mpris2 dataengine. Now the mpris2 engine handles those shortcuts and
delegates to the currently active application.

REVIEW: 122450

[1] http://commits.kde.org/plasma-workspace/409fdfab03e7a0b5a61e7796e1fba85b143eab3b