Organize tracks / Guess tags presets persisted properly.
FilenameLayoutWidget is reworked to fix an old bug regarding the presets
not being saved properly, that I was also able to reproduce.

Logic is not a bit more "linear". Presets are shared between "Organize tracks"
and "Guess tags" dialog. If this is not what is wanted, it can be easily changed,
just let me know. Presets are saved as soon as they are added, updated or deleted.
It is not anymore necessary to accept the dialog to save eventual changes.

The preset used is not saved with the marker "#DELIM#selected" anymore.
Presets are just presets. The one used is matched on the fly using the custom
preset field. If a preset matches the custom preset field, it's considered
selected. Users can of course create custom presets as before.
While presets are shared between the two dialogs, custom presets are not.

Add/Update/Remove work properly. Clicking on a preset name in the combobox
resets the custom pattern to the preset value, this is a new feature,
before it was not possible.

It's a bit difficult to explain, but UX is more natural, just give it a try.

A lot of testing was done (by Stefano), like:
- adding/updating/removing presets.
- if settings are persisted and shared when dialogs are cancelled or accepted.
- if custom patterns are persisted and not shared.
- resetting a modified pattern to its defualt.
- cancelling dialogs, modified presets are kept but custom patterns are not.

REVIEW: 128565
BUG: 226144