Fix MPRIS2 DesktopEntry value
According to the MPRIS2 spec, the DesktopEntry value should be
"the basename of the .desktop file" [1], not the desktop-file-id. So
for amarok it should be "amarok" and not "kde4-amarok", regardless of
where the desktop file is installed. This fixes displaying MPRIS
controls in the Amarok taskbar tooltip in Plasma 5.7

[1] https://specifications.freedesktop.org/mpris-spec/latest/Media_Player.html#Property:DesktopEntry

Testing done by Antonio:

Player controls are displayed again in the Amarok taskbar tooltip in Plasma 5.7.

REVIEW: 128418
BUG: 365275
FIXED-IN: 2.9