Fix Gpodder credential service without kwallet
Before to fix this bug, the gpodder plugin had problems with saving username and password in plain text.

There were two problems:

1) A useless space for the variable m_ignoreWallet was saved in the config file
2) The saving of the credential in plain text where made in the wrong way.
Now it is called the function askAboutMissingKWallet(), that make the use see a dialog windows where is asked to save the credential in plain text.
Credential are now saved in the textDialogYes slot.

REVIEW: 122797