Transcoding: discarding cover art when unsupported
ffmpeg doesn't seem to be able to write the cover art reliably
to output files in m4a, wma, opus, ogg and flac formats.
In some cases transcoding files, in others the cover art
is encoded but can't be read back. For such formats,
the video track that contains the cover art is now discarded.
It was already discarded for opus and ogg.

mp3 is supported, the cover art is now copied unchanged.