Port core application away from kdelibs4
Summary:
This patch ports the core app away from kdelibs4support to pure qt5/kf5. Kdelibs4support classes that don't have an equivalent in qt5 or kf5 like KSharedPtr or KHBox were replaced by new ones (AmarokSharedPtr and BoxWidget).
There are also a lot of small fixes and changes in this patch, which I unfortunately cannot break off into their own revisions, because I lost my git history to a faulty hard drive and my only backup was months old. Sorry for that.
I also temporarily deactivated the tests. They get ported and reactivated in a later patch to keep the scope of this at least somewhat manageable. Also in later patches are a new qml context area with the most important applets, some services and importers. It's all finished and ready for merge as soon as this patch lands.

Reviewers: #amarok, markey

Reviewed By: #amarok, markey

Subscribers: heikobecker, asturmlechner, stikonas, kpiwowarski, markey, spettini, #amarok

Tags: #amarok

Differential Revision: https://phabricator.kde.org/D10526