New context area
Summary: This patch adds a new context area and some of the most important applets for it. The context and applets are written in QML and integrated with a QQuickWidget subclass. Applets consist of a KPackage with their QML files and optionally a QML plugin written in C++ to interface with the core application. It's not very pretty yet but I think this is a good start worth sharing.

Reviewers: #amarok

Subscribers: #amarok

Tags: #amarok

Differential Revision: https://phabricator.kde.org/D10708