Make Webkit optional
Summary:
Qt4Webkit is old, utterly unmaintained and insecure. Quite a number
of distros have abandoned it or are in the process of doing so.
If the option is disabled the Wikipedia and Info context applets
aren't available (as well as the Upcoming events one in combination
with LastFM).

Test Plan: Successfully built with -DWITH_QTWEBKIT=TRUE/FALSE

Reviewers: #amarok, markey

Reviewed By: #amarok, markey

Subscribers: markey

Tags: #amarok

Differential Revision: https://phabricator.kde.org/D10923