Import FindMySQL from kdb
Looks like the author is the same and the version from kdb is a newer
one. Seeing that kdb builds on build.kde.org with the SUSEQt5.9
and FreeBDS images, this wiill hopefully fix Amarok's build there as
well.
Also rename it from MySQLAmarok to MySQL which a) avoids the
impression that you need somehow special customized MySQL version
and b) avoids the awkwardness with differently named foo_FOUND
variables. Furthermore move the find_package(Threads) call to the
main CMakeLists.txt as it's needed in other places.
There's still room for future improvements, but there are quite a
number of other things to attend to.