More interface stuttering work
Summary:
This includes the C++ implementation of the lyrics functionality, which is now async. Also includes a fix for SqlAlbumTest, which failed since my last stuttering revision.
The biggest change though is in CollectionTreeItemModelBase. It now loads an album's tracks in a Threadweaver job and CollectionSortFilterProxyModel filters unloaded albums. This makes the CollectionTreeView seem to grow when expanding an artist with many albums like "Various Artists" instead of freezing for a while depending on your collection size. The CollectionTreeView must be set to display album years for this to take effect.

Reviewers: #amarok, markey

Reviewed By: #amarok, markey

Subscribers: #amarok

Tags: #amarok

Differential Revision: https://phabricator.kde.org/D11968