[TagDialog] Really enable "Open in filemanager" button for local files
`QUrl::QUrl()` expects a full Url including the scheme (unlike
`KUrl::KUrl()` that supports local paths too).
But for local files, `urlString` contains the local path without a
scheme ("file://").

The problem here is that `QUrl` cannot detect that the Url is actually
pointing to a local file if the scheme is missing (according to the Qt
docs, a URL is a local file path if the scheme is "file"), so
`QUrl::isLocalFile()` was always `false` and the "Open in filemanager"
button never got enabled.

So use `QUrl::fromUserInput()` instead to construct the `QUrl`, which
also supports local paths.

Also, to avoid showing the "file://" scheme in the "Location" text field
now, pass `QUrl::PreferLocalFile` to `QUrl::toDisplayString()` as
suggested in the KUrl::pathOrUrl() docs.

Differential Revision: https://phabricator.kde.org/D24043