Register Play/Pause as global shortcut again
Commit 0b2b243b removed the call to KGlobalAccel::setGlobalShortcut()
because it registered Space as global shortcut, which of course is a bad
idea.
But that made it impossible to set up a global shortcut at all (and made
an existing one have no effect).

To fix this, add it back and register an empty key, like the KDE4 code
does.

BUG: 373590
Differential Revision: https://phabricator.kde.org/D24381