Fix activeTrackChanged connect in playlist navigators
The m_model contains a Pointer to an GroupProxy* which was created by
PlaylistModelStack. This cannot be qobject_cast'ed to Playlist::Model* and
thus the connect is to activeTrackChanged is always done against a NULL
pointer instead of the Playlist::Model object.

To actually connect it, the Playlist::Model has to be retrieved from the
Playlist::ModelStack::instance()->bottom() instead of The::playlist()
[m_model].

This broke the cleanup of played entries in dynamic playlist. Also the
repeating tracks didn't play the selected track again but the one which
was selected when this navigator was started.

Fixes: d78f666e58de ("Port core application away from kdelibs4")
BUG: 418643
BUG: 389138