Make include compatible with mariadb
When using mariadb (I still had to set a few things in the CMake
cache manually), the mysql / mariadb include path is added to the
include paths, so no need to include the mysql/ before the header
name. Omitting it allows to get mysql.h from either
[...]/include/mysql/ or [...]/include/mariadb/.

Note that using mariadb seems to be mandatory in Kubuntu 20.04
because the MySQL version it ships does not include libmysqld
anymore.