Fix Increase/Decrease Volume shortcuts
`EngineController::increaseVolume( int ticks )` and
`EngineController::decreaseVolume( int ticks )` take an argument that
specifies the amount by which to increase/decrease the volume.
But it gets called with 0 as argument when pressing the shortcut key, so
the volume was not changed at all.

To make the keyboard shortcuts work again, create new slots that don't
take arguments and call the previous ones with the default value.

Differential Revision: https://phabricator.kde.org/D24413