Remove button to move to the system sound settings
This pointed to (what is now) phononsettings before and currently
points to the kcm from plasma-pa. While the settings obviously might
affect Amarok, the configure dialog is about configuring Amarok
itself and I wouldn't expect a button to change the settings for my
network configuration either, for example. Furthermore this avoids
an optional runtime dependency on pulseaudio and plasma-pa, which
is probably only available on Linux/*BSD.