Fix crash in PodcastSettingsDialog initialization.
Although a box including an Apply button was created, it was not actually set
as the button box for the dialog, and trying to access an Apply button through
buttonBox() thus always resulted in null.
Additionally, make the UI work by converting the setup to actually use the
functionality provided by KPageDialog. Also remove the connection of accepted
and rejected signals, as according to API docs, they are created automatically.

BUG: 393080