Allow copying current track details by clicking.
With text copying from QML Labels being surprisingly hard
(https://bugreports.qt.io/browse/QTBUG-14077), this might be the cleanest
way to implement at least some kind of copy functionality.

A hint of the action is provided with a small copy action icon on hover,
and its opacity is animated on click.

BUG: 283457