Fix first & last played test.
It was previously based on fact that the deprecated QDateTime::toTime_t()
returned -1 cast to an unsigned integer when the QDateTime was invalid.
According to Qt docs, "The behavior for this function [toSecsSinceEpoch]
is undefined if the datetime stored in this object is not valid", so
maybe it is just best to use toString, which has a defined behaviour
of "If the datetime is invalid, an empty string will be returned.".