Fix a potential bug in saved playlist extensions.
Uncovered by the tests that were previously disabled. Apparently
QUrl::fromUserInput behaviour changed between Qt4 and Qt5. In Qt5, with
default options it only returns a path if such local file exists, causing
problems anywhere where a new file was being created. I went through the
uses of the fromUserInput call, and it seems that this was the only place
where it was used on possibly non-existant files.