Remove multihash'y QHash usages
The MetaTagLib cases:
Inserting instead of uniting TagHelper->tags() is straightforward; the possible
tags at that point don't seem to intersect.

The case of combining read and guessed fields is potentially a bit more complicated,
which was noted in code.

In AlbumNavigator: Multiness shouldn't have any effect here

In RandomAlbumNavigator: m_itemsPerAlbum is operated only in AlbumNavigator, and
only with insert and [] operator, so there shouldn't be any multi keys here, either.

In CollectrionTreeItemModel: The structure was already QMultiMap, insertMulti should
equal insert here.