Ignore non-wikipedia links, force monobook skin on opened links
External links are ignored silently now; nothing good comes from trying to open anything
else than wikipedia in the wikipedia applet. Maybe a notification of this would be even
better, but maybe later.

Additionally, edit the opened urls to force the same monobook skin that is used when the
applet automatically opens the next page, and show a busy indicator when opening links.