Change systray icon to use symbolic version
Not having "-symbolic" in name prevents showing of symbolic icon on KF6 Plasma systray.
Looking at various bugreports and code changes done in Yakuake and other applications,
this is apparently ok way to fix this, as it should fall back to icon without symbolic
suffix if none is found.

BUG: 485748

CCBUG: 477228