Fix audiocd advanced settings
Additionally, removes one usage of deprecated KF functions.
The search path is based on my openSUSE system. Probably the correct library
will be found with plain kcm_audiocd.so on some systems, and maybe it is even
possible that some substrings of the full path match the location on some
systems, too. However, it is likely that some other paths also exist for the
library on various distributions, but maybe bug reports will tell about those.

Although there didn't seem to be a report that the advanced settings wasn't
working at all.