Improve encoding when exporting playlist files
Try to avoid excessive escaping local paths when exporting playlist files, and take
one step more to ensure that remote URLs are encoded sensibly.

I tested exporting to each playlist format and both Amarok and VLC seemed to be able
to open each of the files and URLs on them (including local files with non-UTF-8
characters and remote streams with spaces in URL etc). Rhythmbox apparently doesn't
support streams in playlists, but loaded the local files OK, too. In theory, e.g.
m3u shouldn't be UTF-8 (m3u8 is a separate file format for that), but at least all
three tested players successfully loaded tracks with paths saved as UTF-8, so it's
probably OK.

I'm definitely not confident this is perfect now, but I'm quite sure it is notably
better than it was before.

BUG: 333745