Actually check if ServiceFactory descendants are initialized
PluginManager tries to init each of them at the end of checkPluginEnabledStates, but
has no way of knowing whether they are initialized already or not. Storage factories
(=database plugins) had the init check already, add it to ServiceFactories, too, and
make clearing active services reset them to non-initialized state.

I'm not totally sure if this should be done for Collection- and ImporterFactories,
but at least with services this fixes a known bug of multiple instances of each web
service appearing in Internet services browser when saving plugin settings, and
likely also some related crashes.