Make scripty happier by using a different percent symbol escaping
Some Internet result searches suggested that double percent should be the way to
escape a percent in gettext, but apparently scripty scripts didn't agree.
Didn't find any info how to do it, so going for HTML escape then. The actual
message is not used anywhere at the moment anyhow, so it doesn't matter that much.