Update Ampache server version check
BUG: 496581

The value of the `<value>` element in the response from a ping request
to the Ampache server changed from a six-digit integer to a dotted
triple in Ampache version 5.0.0.  This screws up the check for whether
to use the new style of authentication string based on a SHA256 hash, or
the old (now ancient) style of authentication string based on an MD5
hash.  It will try and login to Ampache server version 5.0.0 or later
using the old style of authentication string and so fail to login
successfully.

Use the value of the `<compatible>` element instead if it exists,
falling back to the `<version>` element if it doesn't.  The
`<compatible>` element was introduced in Ampache 3.5.2 and has the
integer value 350001.  This indicates a version older than the actual
version, but it is good enough to determine that the new style of
authentication string should be used.

Also correct the "Version Older than 35001" debug message that has a
missing digit (it should be 350001).

Signed-off-by: Ian Abbott <ian@abbott.org>