Preserve track ordering also when drag-dropping from collection browser
To achieve this, currentOrderTrackLessThan needs to be propagated quite a bit further
so it can be used it when processing dropped track list in PlaylistModel.cpp.

Maybe not the cleanest possible solution, but the required changes are not that
extensive, except for testplaylistmodels, during whose compilation a bunch of related
source files needed to be included.

BUG: 180404