Re-enable at least some cue file functionality
This probably got accidentally disabled in the early '10's when work was
done on removing UI stuttering with MetaProxy usage or so. The autotests
for cue file support have been there & passed all this time. The functionality
was just never invoked.

Still doesn't work if the file is not yet resolved; that is, the audio file
with cue available is added to playlist by dragging the file from external
application, or context view. Will maybe fix later.