Prevent concurrent scan result processings from taking place
Although the scan job itself was already guarded, multiple scan result processings
could be taking place simultaneously. In this case, the same m_directories and
other AbstractScanResultProcessor member variables were used at different points,
and the other result processing was likely to empty those when finishing, causing
the other process to assume that the m_directories was empty all along, and mark
all tracks as missing from collection.

At simplest, if one triggered a new collection refresh while results from the
previous one were still being processed, all tracks were temporarily removed from
the collection. Or, if one added a track to a folder in collection, and after the
watcher had noticed it and started a scan, added another track, this caused the
collection scan to list the tracks in the directory twice and cause a false
warning of all tracks in the folder being duplicated.