Avoid crash when changing database settings
Delaying the execution during updateSettings() (which synchronous KMessageBox does)
causes the Amarok2ConfigDialog destructor to be run before the updateSettings()
finishes (as Qt::WA_DeleteOnClose is set for the dialog), further causing the
remaining config pages to be already deleted and leading to segfault. Avoid by
opening the notification box outside the thread.