Fix example permission grant command in database settings
MariaDB requires a domain (at least if there are anonymous users in db).
Adding domain 'localhost' makes everything work even if there are those
left. If not, wildcard '%' works. Not possible to know from which address
the user connects to the db, so assume that remote connectors have those
anonymous users disabled.

BUG: 386004