Fix continuing playback after TimecodedTracks
Pass playingWillContinue = true to stop, so playlist will not be reset
and the music can continue.

I definitely was not aware of this feature where one can create a track
from two bookmarks and Amarok plays the part between the bookmarks.
Additionally, this fixes continuing playback to next track in cue file.

BUG: 270003